refactor: move AgendaScreenEvent to separate file
