+++ androidApp/src/main/java/com/example/tasky/android/agenda/presentation/agenda/screen/AgendaViewModel.kt
@@ -10,6 +10,7 @@ import com.example.tasky.agenda.domain.model.Event
 import com.example.tasky.agenda.domain.model.RemindAtType
 import com.example.tasky.agenda.domain.model.Reminder
 import com.example.tasky.agenda.domain.model.Task
+import com.example.tasky.android.agenda.presentation.agenda.model.AgendaScreenEvent
 import com.example.tasky.android.agenda.presentation.agenda.model.AgendaScreenState
 import com.example.tasky.auth.domain.IAuthRepository
 import com.example.tasky.auth.domain.ISessionManager
